feat: InputView 클래스로부터 유효성 검사 클래스 추출
* docs: README 리팩토링 목록 - 완료 사항 기록

* feat: 기존의 InputView 클래스로부터 InputViewValidator 클래스 추출

* feat: 기존의 OutputMessage 클래스로부터 InputMessage 클래스 추출

* chore: view/input 패키지를 view/reader 패키지로 이름 변경
