+++ README.md
@@ -49,6 +49,7 @@
 - [x] RandomNumberPicker 클래스에서 사용 중인 매직넘버 분리
 - [x] 클래스 사용 목적 별로 패키지 분리 및 이동
 - [x] DI 관리를 위한 AppConfig 클래스 생성
+- [x] InputView 유효성 검사 메소드 추출. 별도의 클래스 생성
 - [ ] 클래스 및 메소드가 의도에 맞게 명명되었는지 확인
 
 <br>
