+++ src/main/java/racingcar/resource/message/InputMessage.java
@@ -1,0 +1,8 @@
+package racingcar.resource.message;
+
+public class InputMessage {
+
+	public static final String INPUT_NAMES_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
+	public static final String INPUT_EXECUTION_COUNT_MESSAGE = "시도할 회수는 몇회인가요?";
+
+}
