+++ src/main/java/racingcar/resource/message/OutputMessage.java
@@ -2,9 +2,6 @@ package racingcar.resource.message;
 
 public class OutputMessage {
 
-    public static final String INPUT_NAMES_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
-    public static final String INPUT_EXECUTION_COUNT_MESSAGE = "시도할 회수는 몇회인가요?";
-
     public static final String OUTPUT_GAME_RESULT_INTRO_MESSAGE = "실행 결과";
     public static final String OUTPUT_GAME_WINNER_FRONT_MESSAGE = "최종 우승자 : ";
