+++ src/main/java/racingcar/view/reader/ConsoleReader.java
@@ -1,4 +1,4 @@
-package racingcar.view.input;
+package racingcar.view.reader;
 
 import camp.nextstep.edu.missionutils.Console;
